Move part of the quickPreview module to a layout
