+++ src/layouts/quickPreview/components/QuickPreviewBanner.tsx
