+++ src/layouts/quickPreview/components/QuickPreviewLayout.tsx
